Reverting to Asyncio to maintain Windows compatibility
The latest version that appears to be compatible with the test suite is pytest-asyncio 0.21.2. It probably has to do with the way fixtures are initialized, which isn't compatible with VCR.py, but I'm tired of it and will simply revert the version. Someday I'll have the patience to find a solution, but today isn't that day.