Changing Institution Models
Country codes can now be optional. This is in response to some changes in the OpenAlex API. Since the country is always known, but not the country code, it's unclear whether this is an API error. In any case, this change is implemented to prevent errors and will not involve significant changes to the app.